Remove mlockall
mlockall is used to prevent ES from swapping to disk. It locks all the
memory requested on startup. The problem is that if there is no memory
available, it will not be able to start. For whatever reason, ES is
crashing and is not able to restart, which is a problem.